requirements.txt
toml==0.10.2
lxml==4.9.3
coloredlogs==15.0.1
tortoise-orm==0.19.3
aerich==0.7.1
aiogram==3.0.0b7
click==8.1.3
aiohttp==3.8.4
StrEnum==0.4.10
APScheduler==3.10.1
cachetools==5.3.1
aiogram-dialog==2.0.0b18
pytz~=2023.3
matplotlib==3.7.2
beautifulsoup4~=4.12.2